Added Enemy
Wajavook Character: The "Wajavook" character is now part of our game and is programmed to use the NavMeshAgent for tracking and following the player character. It maintains focus on the player's location by using transform.LookAt(target). Physical properties are managed through the Rigidbody component, and animations are handled by the Animator, which can trigger the "StartCrawl" animation.

Wulfvook Character: The "Wulfvook" character is another new addition with the same functionality as the "Wajavook." It uses the NavMeshAgent for navigation, faces the player using transform.LookAt(target), and is equipped with Rigidbody and Animator components for character control and animations.